LSS 119763 Refactor JWT generation in Configuration
