+++ modules/claims_evidence_api/lib/claims_evidence_api/jwt_generator.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'jwt'
+require 'common/client/jwt_generator'
 
 module ClaimsEvidenceApi
   # Usage:
@@ -13,60 +13,22 @@ module ClaimsEvidenceApi
   # Use token:
   # > curl -X GET https://claimevidence-api-test.dev.bip.va.gov/api/v1/rest/swagger-ui.html \
   # > -- 'Authentication: Bearer [TOKEN]'
-  class JwtGenerator
-    # Issuer assigned by Claim Evidence API team
-    ISSUER = 'VAGOV'
+  class JwtGenerator < Common::Client::JwtGenerator
     # VBMS user logged in to the application; if no user interaction needs to be a system user
     USER_ID = 'VAGOVSYSACCT'
     # Station for above user
     STATION_ID = '283'
-    # Number of seconds for which the JWT is valid; should be limited to 15 minutes or less
-    VALIDITY_LENGTH = 15.minutes
-    # Algorithm used to encode and decode the JWT
-    ALGORITHM = 'HS256'
-
-    # static method
-    # @see #encode_jwt
-    def self.encode_jwt
-      new.encode_jwt
-    end
-
-    attr_reader :_headers, :_payload
-
-    # Returns a JWT token for use in Bearer auth
-    def encode_jwt
-      JWT.encode(payload, private_key, ALGORITHM, headers)
-    end
 
     private
 
-    # Returns the headers for the JWT token
-    def headers
-      @_headers = { typ: 'JWT', alg: ALGORITHM }
-    end
-
     # Returns the payload for the JWT token
     def payload
-      @_payload = {
-        jti: SecureRandom.uuid, # random id to identify a unique JWT
-        iat: created_time.to_i,
-        exp: expiration_time.to_i,
-        iss: ISSUER,
-        #  applicationID MUST be the same as the issuer for tracking purposes
+      super.merge({
+        # applicationID MUST be the same as the issuer for tracking purposes
         applicationID: ISSUER,
         userID: USER_ID,
         stationID: STATION_ID
-      }
-    end
-
-    # set the token expiration date
-    def expiration_time
-      created_time + VALIDITY_LENGTH
-    end
-
-    # set the token created time
-    def created_time
-      @created_time ||= Time.zone.now
+      })
     end
 
     # retrieve the secret from settings
